scripts/requirements.txt
discord.py
PyGithub
